+++ test/has_one_has_many_test.rb
@@ -16,12 +16,10 @@ end
 describe "check associations" do
   before(:all) do
     path_to_conf = File.dirname(File.expand_path(__FILE__)) + "/redis.conf"
-    $redis_pid = spawn 'redis-server ' + path_to_conf, :out=>"/dev/null"
-    sleep(1)
-    puts 'started - ' + $redis_pid.to_s
+    $redis_pid = spawn 'redis-server ' + path_to_conf, :out => "/dev/null"
+    sleep(0.3) # must be some delay otherwise "Connection refused - Unable to connect to Redis"
     path_to_socket = File.dirname(File.expand_path(__FILE__)) + "/../redis.sock"
-    puts 'path_to_socket - ' + path_to_socket.inspect
-    $redis = Redis.new(:host => 'localhost', :path => path_to_socket)#:port => 6379)
+    $redis = Redis.new(:host => 'localhost', :path => path_to_socket)
   end
   
   before(:each) do
@@ -33,10 +31,7 @@ describe "check associations" do
   end
 
   after(:all) do
-    puts 'finish - ' + $redis_pid.to_s
-    if $redis_pid
-      Process.kill 9, $redis_pid.to_i
-    end
+    Process.kill 9, $redis_pid.to_i if $redis_pid
   end
 
   it "should save associations properly" do
